+++ src/r_main.c
@@ -308,8 +308,7 @@ static void R_InitTextureMapping(void)
 {
     // Use tangent table to generate viewangletox:
     //  viewangletox will give the next greatest x after the view angle.
-    const fixed_t   limit = finetangent[FINEANGLES / 4 + (((vid_widescreen ? r_fov : r_fov_default)
-                            + WIDEFOVDELTA) * FINEANGLES / 360) / 2];
+    const fixed_t   limit = finetangent[FINEANGLES / 4 + ((r_fov + WIDEFOVDELTA) * FINEANGLES / 360) / 2];
 
     // Calc focallength so field of view angles covers SCREENWIDTH.
     const fixed_t   focallength = FixedDiv(centerxfrac, limit);
@@ -350,8 +349,8 @@ static void R_InitTextureMapping(void)
 //
 void R_InitLightTables(void)
 {
-    const int width = (FixedMul(SCREENWIDTH, FixedDiv(FRACUNIT, finetangent[FINEANGLES / 4 + ((vid_widescreen ?
-                      r_fov : r_fov_default) + WIDEFOVDELTA) * FINEANGLES / 360 / 2])) + 1) / 2 * FRACUNIT;
+    const int width = (FixedMul(SCREENWIDTH, FixedDiv(FRACUNIT, finetangent[FINEANGLES / 4
+                      + (r_fov + WIDEFOVDELTA) * FINEANGLES / 360 / 2])) + 1) / 2 * FRACUNIT;
 
     // Calculate the light levels to use for each level/distance combination.
     for (int i = 0; i < LIGHTLEVELS; i++)
@@ -409,7 +408,7 @@ void R_ExecuteSetViewSize(void)
 
     centerx = viewwidth / 2;
     centerxfrac = centerx << FRACBITS;
-    fovscale = finetangent[FINEANGLES / 4 + ((vid_widescreen ? r_fov : r_fov_default) + WIDEFOVDELTA) * FINEANGLES / 360 / 2];
+    fovscale = finetangent[FINEANGLES / 4 + (r_fov + WIDEFOVDELTA) * FINEANGLES / 360 / 2];
     projection = FixedDiv(centerxfrac, fovscale);
 
     R_InitBuffer();